+++ src/hooks/GJBaseGameLayer.cpp
@@ -1,0 +1,10 @@
+#include "../hooks/GJBaseGameLayer.hpp"
+#include "../hooks/DTPlayLayer.hpp"
+
+void DTGJBaseGameLayer::checkpointActivated(CheckpointGameObject* checkpt) {
+    GJBaseGameLayer::checkpointActivated(checkpt);
+
+    if (static_cast<GJBaseGameLayer*>(this) == PlayLayer::get()){
+        static_cast<DTPlayLayer*>(PlayLayer::get())->m_fields->currentRun.end++;
+    }
+}