+++ src/hooks/GJBaseGameLayer.hpp
@@ -1,0 +1,7 @@
+#include <Geode/modify/GJBaseGameLayer.hpp>
+
+using namespace geode::prelude;
+
+class $modify(DTGJBaseGameLayer, GJBaseGameLayer) {
+    void checkpointActivated(CheckpointGameObject* checkpt);
+};